// packages/rath-client/src/pages/dataConnection/database/interfaces.ts
export type SupportedDatabaseType = (
    | 'postgres'
    | 'clickhouse'
    | 'mysql'
    | 'mariadb'
    | 'sqlite'
    | 'duckdb'
    | 'mongodb'
    | 'clickzetta'
    | 'kylin'
    | 'oracle'
    | 'doris'
    | 'impala'
    | 'awsathena'
    | 'redshift'
    | 'sparksql'
    | 'hive'
    | 'sqlserver'
    | 'drill'
    | 'druid'
    | 'snowflake'
    | 'bigquery'
    | 'demo'
);

export type TableColInfo = {
    key: string;
    colIndex: number;
    dataType: null | string;
};

export type TableRowData = (number | string)[];

export type TableInfo = {
    name: string;
    meta: TableColInfo[];
};